% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_genomeAnnotation.R
\name{create_genomeAnnotation}
\alias{create_genomeAnnotation}
\title{Create genomeAnnotation for ArchR analysis}
\usage{
create_genomeAnnotation(
  BSgenome = NULL,
  geneAnnotation = NULL,
  blacklistBed = NULL,
  header = FALSE
)
}
\arguments{
\item{BSgenome}{An object of \link[BSgenome:BSgenome-class]{BSgenome::BSgenome}.}

\item{geneAnnotation}{A object of SimpleList with three sublist neamed as
"genes", "exons", and "TSS", such as an object outputted by the
\code{\link[=create_geneAnnotation]{create_geneAnnotation()}} function. TSS}

\item{blacklistBed}{A character(1), tab-delimited BED file without a header
line.}

\item{header}{A logical (1), indicating whether the file contains the
names of the variables as its first line.}
}
\value{
An object of SimpleList with three elements as described below.
\describe{
\item{genome}{the name of a BSgenome package}
\item{chromSizes}{A GRanges storing the chromosomal sizes}
\item{blacklist}{A GRanges storing a blacklist for genomic regions to
be excluded}
}
}
\description{
Create a genomeAnnotation object for customize ArchR analysis
}
\examples{
library("txdbmaker")
library("BSgenome.Hsapiens.UCSC.hg38")
gtf <- system.file("extdata", "example.gtf.gz",
                   package = "GenomePal",
                   mustWork = TRUE)
geneID_symbol <- get_geneID_symbol(gtf = gtf)

chrominfo <- data.frame(chrom = c('21','MT'),
                        length=c(46709983, 16569))
TxDb <- makeTxDbFromGFF(file = gtf,
                        format = "gtf",
                        chrominfo = chrominfo,
                        organism = "Homo sapiens",
                        taxonomyId = 9606,
                        circ_seqs = "MT")

## Or using seqinfo of aBSgenome
# library("BSgenome.Hsapiens.UCSC.hg38")
# seqlevelsStyle(BSgenome.Hsapiens.UCSC.hg38) <- "Ensembl"
# seqinfo_hg38 <- seqinfo(BSgenome.Hsapiens.UCSC.hg38)
# seqinfo_hg38 <- seqinfo_hg38[c("21", "MT")]
# TxDb <- makeTxDbFromGFF(file = gtf,
#                        format = "gtf",
#                        chrominfo = seqinfo_hg38,
#                        organism = "Homo sapiens",
#                        taxonomyId = 9606)
geneAnno <- create_geneAnnotation(TxDb = TxDb,
                                  geneID_symbol = geneID_symbol,
                                  filterChr = "MT")

seqlevelsStyle(BSgenome.Hsapiens.UCSC.hg38) <- "ENSEMBL"
genomeAnno <- create_genomeAnnotation(BSgenome = BSgenome.Hsapiens.UCSC.hg38,
                                      geneAnnotation = geneAnno,
                                      blacklistBed = NULL,
                                      header = FALSE)

}
