% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomePal-package.R
\docType{package}
\name{GenomePal-package}
\alias{GenomePal}
\alias{GenomePal-package}
\title{GenomePal: Create Genome Annotation}
\description{
Create a BSgenome package from the genome sequences in a fasta file, and create OrgDb, geneAnnotation, and genomeAnnotation objects from a BSgenome, EnsDb, and GTF, which can be used for customized ArchR-based analyses.
}
\author{
\strong{Maintainer}: Haibo Liu \email{haibo.liu@umassmed.edu} (\href{https://orcid.org/0000-0002-4213-2883}{ORCID})

}
\keyword{internal}
