% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_id_symbol.R
\name{get_geneID_symbol}
\alias{get_geneID_symbol}
\title{Extract gene identifiers and symbols from GTF}
\usage{
get_geneID_symbol(gtf = NULL, unique_symbol = FALSE)
}
\arguments{
\item{gtf}{A character(1), specifying a path to an Ensembl GTF file. Many
UCSC GTF files are malformatted, thus should not be used.}

\item{unique_symbol}{A logical(1), whether only return genes with unique
symbols. Default is FALSE.}
}
\value{
A data frame with two columns: \code{gene_id} and \code{symbol}.
}
\description{
Extract gene identifiers and symbols from GTF
}
\examples{
gtf <- system.file("extdata", "example.gtf.gz",
                   package = "GenomePal",
                   mustWork = TRUE)
geneID_symbol <- get_geneID_symbol(gtf = gtf)

}
