% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_BSgenome_from_fasta.R
\name{make_BSgenome_pkg}
\alias{make_BSgenome_pkg}
\title{Create a BSgenome package}
\usage{
make_BSgenome_pkg(
  genome_fasta = NULL,
  dest_dir = tempdir(),
  prefix = c("", "chr"),
  latin_name = NULL,
  common_name = NULL,
  genome_version = NULL,
  seed_file_name = NULL,
  fasta_url = NULL,
  release_date = "August 2007",
  source = "Ensembl",
  version = "1.0.0"
)
}
\arguments{
\item{genome_fasta}{A character(1) specifying a path for a multi-fasta file
for a reference genome. It can be a gzip-compressed or uncompressed fasta
file.}

\item{dest_dir}{A character(1) specifying a path for output split, compressed
fasta files}

\item{prefix}{A character(1), specifying chromosome prefix. It can be an
empty string or "chr". Default is "".}

\item{latin_name}{A character(1), specifying the Latin name of the species of
the reference genome, such as "Homo sapiens" for the human.}

\item{common_name}{A character(1), specifying the common name of the species
of the reference genome, such as "Human".}

\item{genome_version}{A character(1), specifying the genome build, such as
"GRCh38".}

\item{seed_file_name}{A character(1), specifying the path to a seed file to
be created for building a BSgenome package.}

\item{fasta_url}{A character(1), specifying the URL from where the reference
genome fasta file is downloaded.}

\item{release_date}{A character(1), specifying the genome assembly release
date, such as "August 2020".}

\item{source}{A character(1), specifying the source of the genome fasta
file, such as "Ensembl".}

\item{version}{A character(1), specifying the version of a BSgenome package
to be build, such as "1.0.0".}
}
\value{
A BSgenome package name. Users have to install the package
to use it.
}
\description{
Starting with a multi-fasta file for a reference genome, a BSgenome package
is created and installed.
}
\examples{
dest_dir <- tempdir()
genome_fasta <- file.path(dest_dir, "toy.example.fa")
in_fasta <- file(genome_fasta, open = "w")
writeLines(c(">chr1\nATCGCTGCGGATGCGG",
         ">chr2\nCCCCCCCCCCCGGAAA",
         ">chrM\nATACTACTGGA"), in_fasta)
close(in_fasta)

make_BSgenome_pkg(
   genome_fasta = genome_fasta,
   dest_dir = file.path(dest_dir, "multifasta"),
   prefix = "none",
   latin_name = "Homo sapiens",
   common_name = "Human",
   genome_version = "GRCh38",
   seed_file_name = file.path(
       dest_dir,
       "human.genome.seed.txt"
   ),
   fasta_url = "http://ftp.ensembl.org/xxx.fa",
   release_date = "August 2007",
   source = "Ensembl",
   version = "1.0.0"
)
}
